LayoutTests/fast/events/onunload-not-on-body-expected.txt

frame "<!--framePath /1/<!--frame2-->-->" - has 1 onunload handler(s)
ALERT: unload
you should only see one unload alert appear.
